requirements.txt
biopython
requests
pandas
scikit-learn
umap-learn
matplotlib
seaborn
scipy
numpy
